requirements.txt
fastapi==0.92.0
python-decouple==3.8
starlette==0.25.0
uvicorn==0.20.0